Annotate peptides with protein ids

peptides which do not have protein assignment drop out

annotatePeptides(
  pepinfo,
  fasta,
  peptide = "peptideSeq",
  prefix = "(([RK])|(^)|(^M))",
  suffix = ""
)

Arguments

pepinfo
- list of peptides - sequence, optional modified sequence, charge state.

fasta
- object as created by readPeptideFasta

peptide
- name of column containing peptide sequences default "peptideSeq"

prefix
- default "(([RK])|(^)|(^M))"

suffix
- default ""

Examples

library(prozor)
library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

file = system.file("extdata/shortfasta.fasta.gz",package = "prozor")

fasta = readPeptideFasta(file = file)
res = annotatePeptides(pepprot[1:20,], fasta)

#> Warning: `funs()` was deprecated in dplyr 0.8.0.
#> Please use a list of either functions or lambdas: 
#> 
#>   # Simple named list: 
#>   list(mean = mean, median = median)
#> 
#>   # Auto named with `tibble::lst()`: 
#>   tibble::lst(mean, median)
#> 
#>   # Using lambdas
#>   list(~ mean(., trim = .2), ~ median(., na.rm = TRUE))

#> Joining, by = "proteinID"

res = annotatePeptides(pepprot[1:20,"peptideSeq"],fasta)

#> Joining, by = "proteinID"

str(res)

#> 'data.frame':	45 obs. of  6 variables:
#>  $ peptideSeq     : chr  "AAAITSDLLESLGR" "AAAITSDLLESLGR" "AAAITSDLLESLGR" "AAAITSDLLESLGR" ...
#>  $ proteinID      : chr  "sp|Q62009|POSTN_MOUSE" "sp|Q62009-3|POSTN_MOUSE" "sp|Q62009-4|POSTN_MOUSE" "sp|Q62009-5|POSTN_MOUSE" ...
#>  $ Offset         : num  254 254 254 254 254 266 266 297 218 585 ...
#>  $ proteinSequence: chr  "MVPLLPLYALLLLFLCDINPANANSYYDKVLAHSRIRGRDQGPNVCALQQILGTKKKYFSSCKNWYQGAICGKKTTVLYECCPGYMRMEGMKGCPAVMPIDHVYGTLGIVG"| __truncated__ "MVPLLPLYALLLLFLCDINPANANSYYDKVLAHSRIRGRDQGPNVCALQQILGTKKKYFSSCKNWYQGAICGKKTTVLYECCPGYMRMEGMKGCPAVMPIDHVYGTLGIVG"| __truncated__ "MVPLLPLYALLLLFLCDINPANANSYYDKVLAHSRIRGRDQGPNVCALQQILGTKKKYFSSCKNWYQGAICGKKTTVLYECCPGYMRMEGMKGCPAVMPIDHVYGTLGIVG"| __truncated__ "MVPLLPLYALLLLFLCDINPANANSYYDKVLAHSRIRGRDQGPNVCALQQILGTKKKYFSSCKNWYQGAICGKKTTVLYECCPGYMRMEGMKGCPAVMPIDHVYGTLGIVG"| __truncated__ ...
#>  $ matched        : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ lengthPeptide  : int  14 14 14 14 14 8 8 9 9 18 ...

res %>% mutate(proteinlength = nchar(proteinSequence)) -> res
res %>% select(proteinID, peptideSeq, proteinlength, Offset, lengthPeptide)

#>                  proteinID               peptideSeq proteinlength Offset
#> 1    sp|Q62009|POSTN_MOUSE           AAAITSDLLESLGR           838    254
#> 2  sp|Q62009-3|POSTN_MOUSE           AAAITSDLLESLGR           810    254
#> 3  sp|Q62009-4|POSTN_MOUSE           AAAITSDLLESLGR           784    254
#> 4  sp|Q62009-5|POSTN_MOUSE           AAAITSDLLESLGR           783    254
#> 5  sp|Q62009-2|POSTN_MOUSE           AAAITSDLLESLGR           811    254
#> 6   tr|Q543K9|Q543K9_MOUSE                 AAAQTLER           289    266
#> 7     sp|P23492|PNPH_MOUSE                 AAAQTLER           289    266
#> 8     sp|P04919|B3AT_MOUSE                AAATLMTER           929    297
#> 9   sp|P04919-2|B3AT_MOUSE                AAATLMTER           850    218
#> 10    sp|P07724|ALBU_MOUSE       AADKDTCFSTEGPNLVTR           608    585
#> 11    sp|P07724|ALBU_MOUSE       AADKDTCFSTEGPNLVTR           608    585
#> 12    sp|P07724|ALBU_MOUSE       AADKDTCFSTEGPNLVTR           608    585
#> 13   sp|P62259|1433E_MOUSE      AAFDDAIAELDTLSEESYK           255    197
#> 14    sp|P97298|PEDF_MOUSE   AAFEWNEEGAGSSPSPGLQPVR           417    359
#> 15  tr|A2A997|A2A997_MOUSE     AALGYNILTQEEAQSVYDAK           580    152
#> 16  tr|A2A997|A2A997_MOUSE     AALGYNILTQEEAQSVYDAK           580    152
#> 17  tr|A2A998|A2A998_MOUSE     AALGYNILTQEEAQSVYDAK           543    108
#> 18  tr|A2A998|A2A998_MOUSE     AALGYNILTQEEAQSVYDAK           543    108
#> 19    sp|Q8K182|CO8A_MOUSE     AALGYNILTQEEAQSVYDAK           587    152
#> 20    sp|Q8K182|CO8A_MOUSE     AALGYNILTQEEAQSVYDAK           587    152
#> 21    sp|P49722|PSA2_MOUSE              AANGVVLATEK           234     40
#> 22    sp|O89020|AFAM_MOUSE                AAPITQYLK           608    206
#> 23  sp|O89020-3|AFAM_MOUSE                AAPITQYLK           611    206
#> 24  sp|O89020-2|AFAM_MOUSE                AAPITQYLK           430    206
#> 25     sp|Q61838|A2M_MOUSE AAPLSLCALTAVDQSVLLLKPEAK          1495    588
#> 26     sp|Q61838|A2M_MOUSE AAPLSLCALTAVDQSVLLLKPEAK          1495    588
#> 27     sp|Q61838|A2M_MOUSE AAPLSLCALTAVDQSVLLLKPEAK          1495    588
#> 28  tr|D3YW52|D3YW52_MOUSE AAPLSLCALTAVDQSVLLLKPEAK          1507    588
#> 29  tr|D3YW52|D3YW52_MOUSE AAPLSLCALTAVDQSVLLLKPEAK          1507    588
#> 30  tr|D3YW52|D3YW52_MOUSE AAPLSLCALTAVDQSVLLLKPEAK          1507    588
#> 31    sp|O89020|AFAM_MOUSE          AAPQLPMEELVSLSK           608    438
#> 32  sp|O89020-3|AFAM_MOUSE          AAPQLPMEELVSLSK           611    438
#> 33  tr|Q9D1B1|Q9D1B1_MOUSE     AAQAAVASYNMGSDSLYYFR           149     51
#> 34  tr|A6ZI44|A6ZI44_MOUSE                 AAQEEYIK           418    377
#> 35  tr|Q9CPQ9|Q9CPQ9_MOUSE                 AAQEEYIK           364    323
#> 36  tr|A6ZI46|A6ZI46_MOUSE                 AAQEEYIK           419    378
#> 37   sp|P05064|ALDOA_MOUSE                 AAQEEYIK           364    323
#> 38  tr|Q3UTR7|Q3UTR7_MOUSE           AAQVAMIANFVGFR           482     99
#> 39    sp|P11859|ANGT_MOUSE           AAQVAMIANFVGFR           477     94
#> 40  tr|Q8BSB7|Q8BSB7_MOUSE    AASGAERPDFCFLEEDPGLCR           235    111
#> 41  tr|A2ARA4|A2ARA4_MOUSE    AASGAERPDFCFLEEDPGLCR           246    104
#> 42  tr|A2ARA0|A2ARA0_MOUSE    AASGAERPDFCFLEEDPGLCR           300    111
#> 43 sp|O54819-2|TFPI1_MOUSE    AASGAERPDFCFLEEDPGLCR           253    111
#> 44  tr|A2ARA3|A2ARA3_MOUSE    AASGAERPDFCFLEEDPGLCR           247    111
#> 45   sp|O54819|TFPI1_MOUSE    AASGAERPDFCFLEEDPGLCR           306    111
#>    lengthPeptide
#> 1             14
#> 2             14
#> 3             14
#> 4             14
#> 5             14
#> 6              8
#> 7              8
#> 8              9
#> 9              9
#> 10            18
#> 11            18
#> 12            18
#> 13            19
#> 14            22
#> 15            20
#> 16            20
#> 17            20
#> 18            20
#> 19            20
#> 20            20
#> 21            11
#> 22             9
#> 23             9
#> 24             9
#> 25            24
#> 26            24
#> 27            24
#> 28            24
#> 29            24
#> 30            24
#> 31            15
#> 32            15
#> 33            20
#> 34             8
#> 35             8
#> 36             8
#> 37             8
#> 38            14
#> 39            14
#> 40            21
#> 41            21
#> 42            21
#> 43            21
#> 44            21
#> 45            21